Move shared code to util file
