+++ lib/compose.js
@@ -50,7 +50,6 @@ function asyncify(f) {
   }
 }
 
-
 // This accepts an array of arrays and outputs a combination of their layers.
 //
 // combineLayers([ [ 1, 2 ], [ 3, 4] ]) // => [
@@ -73,7 +72,12 @@ const combineLayers = x =>
     acc.map(chain =>
       group.map((item) =>
         chain.concat(item))
-    ).flat(), [[]])
+    ).reduce((acc, val) =>
+      // alternative to `Array.prototype.flat()`
+      acc.concat(val), []
+    ),
+    [[]]
+  )
 
 module.exports = function (ary, wrap) {
   if(!wrap) wrap = function (e) { return e }
