+++ plugins/net.js
@@ -1,65 +1,25 @@
+const toPull = require('stream-to-pull-stream')
+const debug = require('debug')('multiserver:net')
+const {
+  getAddress,
+  getRandomPort,
+  protocolToAddress
+} = require('../lib/util.js')
+
 var net
-var os
 try {
   net = require('net')
-  os = require('os')
 } catch (_) {
-  // This only throws in browsers because they don't have access to the Node
-  // net library, which is safe to ignore because they shouldn't be running
-  // any methods that require the net library. Maybe we should be setting a
-  // flag somewhere rather than checking whether `net == null`?
+  // Uncaught because this should work in a browser.
 }
 
-var toPull = require('stream-to-pull-stream')
-var scopes = require('multiserver-scopes')
-var debug = require('debug')('multiserver:net')
-
-const isString = (s) => 'string' == typeof s
-const toAddress = (host, port) => ['net', host, port ].join(':')
+const toAddress = protocolToAddress('net')
 
 function toDuplex (str) {
   var stream = toPull.duplex(str)
   stream.address = toAddress(str.remoteAddress, str.remotePort)
   return stream
 }
-
-const metaAddresses = [
-  '0.0.0.0',
-  '::'
-]
-
-const getAddress = (host, scope) => {
-  if (scope === 'device') {
-    return 'localhost' // legacy
-  }
-
-  if (typeof host === 'string' && metaAddresses.includes(host) === false) {
-    return host
-  }
-
-  return getNetworkAddresses({
-    internal: (scope === 'device'),
-    family: (host === '0.0.0.0' ? 'IPv4' : null)
-  })
-}
-
-const getNetworkAddresses = ({ internal, family } = {}) =>
-  Object.values(os.networkInterfaces())
-    // Flatten
-    .reduce((acc, val) => acc.concat(val), [])
-    // Filter `internal`
-    .filter(item => internal == null || item.internal === internal)
-    // Filter `family`
-    .filter(item => family == null || item.family === family)
-    // Filter scoped IPv6 addresses, which don't play nicely with Node.js
-    .filter(item => item.scopeid == null || item.scopeid === 0)
-    // Only return the address.
-    .map(item => item.address)
-
-// Choose a dynamic port between 49152 and 65535
-// https://en.wikipedia.org/wiki/List_of_TCP_and_UDP_port_numbers#Dynamic,_private_or_ephemeral_ports
-const getRandomPort = () =>
-  Math.floor(49152 + (65535 - 49152 + 1) * Math.random())
 
 module.exports = ({ scope = 'device', host, port, external, allowHalfOpen, pauseOnConnect }) => {
   // Arguments are `scope` and `external` plus selected options for
